@@ -3,20 +3,24 @@
   "version": "1.5.1",
   "description": "A Vue.js syntax highlighter that reads the source code you give it and highlights it, for Humans.",
   "author": "Antoni Andre <antoniandre.web@gmail.com>",
+  "homepage": "https://antoniandre.github.io/simple-syntax-highlighter",
   "repository": "https://github.com/antoniandre/simple-syntax-highlighter",
-  "main": "dist/sshpre.common.js",
-  "unpkg": "dist/sshpre.umd.min.js",
+  "license": "MIT",
+  "funding": "https://github.com/sponsors/antoniandre",
+  "main": "dist/sshpre.cjs.js",
+  "unpkg": "dist/sshpre.umd.js",
+  "jsdelivr": "dist/sshpre.umd.js",
   "module": "./dist/sshpre.es.js",
   "files": [
     "dist/"
   ],
   "exports": {
-    "./dist/sshpre.css": "./dist/sshpre.css",
-    "./dist/sshpre.cjs.js": "./dist/sshpre.cjs.js",
     ".": {
       "import": "./dist/sshpre.es.js",
       "require": "./dist/sshpre.umd.js"
-    }
+    },
+    "./package.json": "./package.json",
+    "./dist/*": "./dist/*"
   },
   "keywords": [
     "syntax highlighter",
@@ -27,7 +31,6 @@
     "vuejs",
     "ui"
   ],
-  "license": "MIT",
   "scripts": {
     "dev": "vite",
     "build": "vite build --base /simple-syntax-highlighter/",
@@ -35,28 +38,25 @@
     "serve": "vite preview --base /simple-syntax-highlighter/"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "^7.17.0",
-    "@vitejs/plugin-vue": "^1.10.2",
+    "@babel/eslint-parser": "^7.19.1",
+    "@vitejs/plugin-vue": "^3.2.0",
+    "@vue/compiler-sfc": "^3.2.47",
     "@vue/eslint-config-standard": "^6.1.0",
-    "autoprefixer": "^10.4.2",
-    "core-js": "^3.21.1",
+    "autoprefixer": "^10.4.13",
     "eslint": "^7.32.0",
-    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^5.2.0",
     "eslint-plugin-vue": "^7.20.0",
+    "postcss": "^8.4.21",
+    "pug": "^3.0.2",
     "rollup-plugin-delete": "^2.0.0",
-    "sass": "^1.49.8",
-    "vite": "^2.8.4",
-    "vite-plugin-pug": "^0.3.0",
-    "vite-plugin-vue2": "^1.9.3",
-    "vue": "^2.6.14",
-    "vue-template-compiler": "^2.6.14",
-    "wave-ui": "^1.49.3"
+    "sass": "^1.58.0",
+    "vite": "^3.2.5",
+    "vue": "^3.2.47",
+    "wave-ui": "^3.0.0"
   },
-  "pnpm": {
-    "overrides": {
-      "postcss@<8.2.13": ">=8.2.13"
-    }
+  "peerDependencies": {
+    "vue": "^2.6.14 || ^3.2.0"
   }
 }
