@@ -1,10 +1,28 @@
-import Vue from 'vue'
-import WaveUI from './plugins/wave-ui'
+import { createApp } from 'vue'
+import WaveUI from 'wave-ui'
+import 'wave-ui/dist/wave-ui.css'
 import App from './app.vue'
 
-Vue.config.productionTip = false
+const app = createApp(App)
 
-new Vue({ // eslint-disable-line no-new
-  WaveUI,
-  render: h => h(App)
-}).$mount('#app')
+app.use(WaveUI, {
+  icons: ['md', 'fa'],
+  iconsLigature: 'material-icons',
+  colors: {
+    light: {
+      primary: '#1b4',
+      secondary: '#666',
+      maintext: '#333',
+      lightgrey: '#eee'
+    },
+    dark: {
+      primary: '#6c8',
+      secondary: '#aaa',
+      maintext: '#999',
+      lightgrey: '#eee'
+    }
+  },
+  theme: 'auto'
+})
+
+app.mount('#app')
