@@ -1,55 +1,75 @@
 <template lang="pug">
 .documentation
-  div.mb3.title3.black Github project
+  w-button.mr1.theme-switch(
+    fixed
+    top
+    right
+    :icon="`material-icons ${$waveui.theme === 'dark' ? 'light_mode' : 'dark_mode'}`"
+    @click="$waveui.switchTheme($waveui.theme === 'light' ? 'dark' : 'light')")
+
+  div.mb3.title3 Github project
   w-flex.mb5(align-center shrink)
     w-icon.mr4.lightgrey(size="40") fab fa-github
     a(href="https://github.com/antoniandre/simple-syntax-highlighter" target="_blank") //github.com/antoniandre/simple-syntax-highlighter
 
   h2
-    a(href="#What-is-it") What is it?
+    a(href="#What-is-it") No chi-chi syntax highlighter for Vue 2 &amp; 3.
     a(name="What-is-it")
-  .title4.mt5.mb3.lh3
-    strong Simple Syntax Highlighter is is a lightweight yet efficient - and very easy to use - syntax highlighter for Vue JS.#[br]
-    | It reads the source code you give it and syntax-highlights it, for Humans.
-  p With its very small size, it is not an editor but a parser.
+  .title4.mt4.lh3
+    ul.checklist.lh2
+      li #[span.wi-check.w-icon.success] Easy to use
+      li #[span.wi-check.w-icon.success] Easy to customize
+      li #[span.wi-check.w-icon.success] Very light weight
+      li #[span.wi-check.w-icon.success] Support multiple languages
+      li #[span.wi-check.w-icon.success] Support both reading and editing
+      li #[span.wi-check.w-icon.success] Built-in copy to clipboard button
+      li #[span.wi-check.w-icon.success] Built-in light and dark modes
 
-  h3.mt8
-    w-icon.ml-1 material-icons chevron_right
-    | Supports Vue 2 &amp; Vue 3
-  div All the features of Simple Syntax Highlighter are working as expected in Vue 3.
+  h3.mt8.ml6.mb2 Supported languages
+  ul.checklist.lh2.ml2.mt0
+    li #[span.wi-check.w-icon.success] XML
+    li #[span.wi-check.w-icon.success] HTML
+    li #[span.wi-check.w-icon.success] HTML Vue JS Template
+    li #[span.wi-check.w-icon.success] Pug
+    li #[span.wi-check.w-icon.success] JavaScript
+    li #[span.wi-check.w-icon.success] JSON
+    li #[span.wi-check.w-icon.success] CSS
+    li #[span.wi-check.w-icon.success] PHP
+    li #[span.wi-check.w-icon.success] MySQL
+    li #[span.wi-cross.w-icon.error] No support for mixed nested languages
 
-  h3.mt8
-    w-icon.ml-1 material-icons chevron_right
-    | Supported languages
-  ul.lh2.ml1
-    li XML
-    li HTML
-    li HTML Vue JS Template
-    li Pug
-    li Javascript
-    li JSON
-    li CSS
-    li PHP
-    li MySQL
-    li No support for nested languages
+
+  h2
+    a(href="#demo") demo - try it!
+    a(name="demo")
+  ssh-pre(language="js" label="Javascript" :dark="$waveui.theme === 'dark'").
+    /**
+     * I can syntax-highlight your JavaScript code.
+     * Try to type some code!
+     */
+
+    import Vue from 'vue'
+
+    const radius = 25 // Radius in pixels.
+    console.log(2 * Math.Pi * radius)
 
   h2
     a(href="#installation") Installation
     a(name="installation")
   p You have two options: #[em NPM] #[strong.mx2 or] #[span.code &lt;script&gt;] tag.
+
   h3.mt8 Via NPM
-  ssh-pre(language="shell").
-    npm i simple-syntax-highlighter # Vue 2.x.
+  .w-flex.align-center.my4.xs-column
+    ssh-pre.my0(language="shell" :dark="$waveui.theme === 'dark'") npm i simple-syntax-highlighter # Vue 3
+    span.mx6 or
+    ssh-pre.my0(language="shell" :dark="$waveui.theme === 'dark'") npm i simple-syntax-highlighter@legacy # Vue 2
 
-  ssh-pre(language="shell").
-    npm i simple-syntax-highlighter@next # Vue 3.
-
-  p.mt2.grey--text.text--darken-1
-    | Edit a working Vue 3 example on Codepen:
+  p.grey--text.text--darken-1
+    | Edit a Vue 3 example on Codepen:
     a.ml2(href="https://codepen.io/antoniandre/pen/jOqgoKR" target="_blank") //codepen.io/antoniandre/pen/jOqgoKR
 
   p.mt12 Then import the component and use it:
-  ssh-pre(language="js" label="Javascript").
+  ssh-pre(language="js" label="Javascript" :dark="$waveui.theme === 'dark'").
     // In your VueJS component.
     import SshPre from 'simple-syntax-highlighter'
     import 'simple-syntax-highlighter/dist/sshpre.css'
@@ -63,7 +83,7 @@
     | In some cases, like in vue-cli or Vite, you might need to preserve the white spaces in the project
     | (if you see all the ssh-pre contents on a single line).#[br]
     | Here is how to preserve whitespaces in Vue. Add this in your vue.config.js file:
-    ssh-pre.my1(language="js").
+    ssh-pre.my1(language="js" :dark="$waveui.theme === 'dark'").
       module.exports = {
         chainWebpack: config =&gt; {
           // Preserve white spaces for ssh-pre component.
@@ -78,7 +98,7 @@
         }
       }
     p Or in Vite, add this in your vite.config.js file:
-    ssh-pre.my1(language="js").
+    ssh-pre.my1(language="js" :dark="$waveui.theme === 'dark'").
       export default defineConfig({
         plugins: [
           createVuePlugin({
@@ -93,13 +113,27 @@
 
   h3.mt8 Via #[span.code &lt;script&gt;] tag
   p Include the Simple Syntax Highlighter script in your document #[span.code &lt;head&gt;] as follows:
-  ssh-pre(language="html" label="HTML").
-    &lt;head&gt;
-      ...
-      &lt;script src="https://unpkg.com/vue"&gt;&lt;/script&gt;
-      &lt;script src="https://unpkg.com/simple-syntax-highlighter"&gt;&lt;/script&gt;
-      &lt;link href="https://unpkg.com/simple-syntax-highlighter/dist/sshpre.css" rel="stylesheet"&gt;
-    &lt;/head&gt;
+
+  w-tabs.my4(:items="2" content-class="pa0")
+    template(#item-title.1) Vue 3
+    template(#item-content.1)
+      ssh-pre.ma0(language="html" :dark="$waveui.theme === 'dark'").
+        &lt;head&gt;
+          ...
+          &lt;script src="https://unpkg.com/vue"&gt;&lt;/script&gt;
+          &lt;script src="https://unpkg.com/vue-cal"&gt;&lt;/script&gt;
+          &lt;link href="https://unpkg.com/vue-cal/dist/vuecal.css" rel="stylesheet"&gt;
+        &lt;/head&gt;
+
+    template(#item-title.2) Vue 2
+    template(#item-content.2)
+      ssh-pre.ma0(language="html" :dark="$waveui.theme === 'dark'").
+        &lt;head&gt;
+          ...
+          &lt;script src="https://unpkg.com/vue@legacy"&gt;&lt;/script&gt;
+          &lt;script src="https://unpkg.com/vue-cal@legacy"&gt;&lt;/script&gt;
+          &lt;link href="https://unpkg.com/vue-cal@legacy/dist/vuecal.css" rel="stylesheet"&gt;
+        &lt;/head&gt;
 
   h2
     a(href="#how-to-use") How to use
@@ -109,7 +143,10 @@
 
   h3.mt8 For any other language than HTML
   w-flex.responsive-block(wrap align-center justify-center)
-    ssh-pre.grow.basis-zero(language="html-vue" label="HTML Vue Template").
+    ssh-pre.grow.basis-zero(
+      language="html-vue"
+      label="HTML Vue Template"
+      :dark="$waveui.theme === 'dark'").
       &lt;ssh-pre language="js" label="Javascript"&gt;
         const i = 3;
         // Some more Javascript content.
@@ -117,7 +154,10 @@
     div.mt8.mx4.text-center.arrow-block
       div renders as:
       w-icon(xl) material-icons arrow_right_alt
-    ssh-pre.grow.basis-zero(language="js" label="Javascript").
+    ssh-pre.grow.basis-zero(
+      language="js"
+      label="Javascript"
+      :dark="$waveui.theme === 'dark'").
       const i = 3;
       // Some more Javascript content.
 
@@ -126,21 +166,27 @@
     strong To use XML/HTML-like languages: you must escape the chevrons: #[code &amp;lt;] for #[code &lt;] and #[code &amp;gt;] for #[code &gt;].
     | #[br]If you want to display an #[code &amp;] character, you need to escape it with: #[code &amp;amp;amp;].
   w-flex.responsive-block(wrap align-center justify-center)
-    ssh-pre.grow.basis-zero(language="html-vue" label="HTML").
+    ssh-pre.grow.basis-zero(
+      language="html-vue"
+      label="HTML"
+      :dark="$waveui.theme === 'dark'").
       &lt;ssh-pre language="html" label="HTML Vue Template"&gt;
-        &amp;amp;lt;html lang="en"&amp;amp;gt;
-          &amp;amp;lt;head&amp;amp;gt;
-            &amp;amp;lt;title&amp;amp;gt;Title&amp;amp;lt;/title&amp;amp;gt;
-          &amp;amp;lt;/head&amp;amp;gt;
-          &amp;amp;lt;body&amp;amp;gt;
+        &amp;lt;html lang="en"&amp;gt;
+          &amp;lt;head&amp;gt;
+            &amp;lt;title&amp;gt;Title&amp;lt;/title&amp;gt;
+          &amp;lt;/head&amp;gt;
+          &amp;lt;body&amp;gt;
             Some content.
-          &amp;amp;lt;/body&amp;amp;gt;
-        &amp;amp;lt;/html&amp;amp;gt;
+          &amp;lt;/body&amp;gt;
+        &amp;lt;/html&amp;gt;
       &lt;/ssh-pre&gt;
     div.mx4.text-center.arrow-block
       div renders as:
       w-icon(xl) material-icons arrow_right_alt
-    ssh-pre.grow.basis-zero(language="html-vue" label="HTML Vue Template").
+    ssh-pre.grow.basis-zero(
+      language="html-vue"
+      label="HTML Vue Template"
+      :dark="$waveui.theme === 'dark'").
       &lt;html lang="en"&gt;
         &lt;head&gt;
           &lt;title&gt;Title&lt;/title&gt;
@@ -165,10 +211,10 @@
       span.code.blue--text.text--darken-1.mx2 Default: #[strong '']
       p One of the supported programming languages to syntax highlight.
     li
-      code reactive
+      code reactive (prior v.3.0)
       span.code.cyan--text.text--darken-3.mx2 {Boolean}
       span.code.blue--text.text--darken-1.mx2 Default: #[strong false]
-      p Makes this ssh-pre box reactive to variable changes.
+      p Makes this ssh-pre box reactive to variable changes (not needed from version 3.0).
     li
       code dark
       span.code.cyan--text.text--darken-3.mx2 {Boolean}
@@ -191,14 +237,22 @@
   .w-flex.wrap.justify-center.mxa.mt2.responsive-block(style="max-width: 650px")
     div.grow.basis-zero.my2
       .text-bold Default
-      ssh-pre(language="html" copy-button).
+      ssh-pre(
+        language="html"
+        copy-button
+        @copied="$waveui.notify('copied to clipboard!', 'success')"
+        :dark="$waveui.theme === 'dark'").
         &lt;body&gt;
           &lt;div id="app"&gt;&lt;/div&gt;
         &lt;/body&gt;
     .spacer.mx1.no-grow
     div.grow.basis-zero.my2
       .text-bold Custom button content
-      ssh-pre(language="html" copy-button)
+      ssh-pre(
+        language="html"
+        copy-button
+        @copied="$waveui.notify('copied to clipboard!', 'success')"
+        :dark="$waveui.theme === 'dark'")
         template(#copy-button)
           w-icon(small color="primary") material-icons content_copy
         | &lt;body&gt;
@@ -209,7 +263,7 @@
     On copy button press, the #[span.code copied] event is emitted containing the copied text.#[br]
     Here is how to use the copy-button slot and the copied event.
 
-  ssh-pre(language="html-vue").
+  ssh-pre(language="html-vue" :dark="$waveui.theme === 'dark'").
     &lt;ssh-pre language="html-vue" copy-button @copied="onCopiedDoSomething"&gt;
       &lt;template #copy-button&gt;
         &lt;i class="icon material-icons"&gt;material-icons content_copy&lt;/i&gt;
@@ -224,7 +278,7 @@
   h3
     a(href="#ex--javascript") javascript
     a(name="ex--javascript")
-  ssh-pre(language="js" label="Javascript").
+  ssh-pre(language="js" label="Javascript" :dark="$waveui.theme === 'dark'").
     /**
      * This is some
      * multiline comments.
@@ -266,7 +320,7 @@
   h3
     a(href="#ex--css") CSS
     a(name="ex--css")
-  ssh-pre(language="css" label="styles.css").
+  ssh-pre(language="css" label="styles.css" :dark="$waveui.theme === 'dark'").
     :root {
       --primary-color: green;
     }
@@ -349,7 +403,7 @@
   h3
     a(href="#ex--html") HTML
     a(name="ex--html")
-  ssh-pre(language="xml" label="index.html").
+  ssh-pre(language="xml" label="index.html" :dark="$waveui.theme === 'dark'").
     &lt;!DOCTYPE html&gt;
     &lt;html lang="en"&gt;
     &lt;head&gt;
@@ -374,7 +428,7 @@
   p.
     Important: to put parenthesis inside quotes, you have to escape them with: #[code="&amp;#40;"]
     for #[code (] and #[code="&amp;#41;"] for #[code )].
-  ssh-pre(language="pug" label="index.pug").
+  ssh-pre(language="pug" label="index.pug" :dark="$waveui.theme === 'dark'").
     //- This is pug.
     div.mb3.title3.black Github project
 
@@ -416,7 +470,7 @@
   h3
     a(href="#ex--json") JSON
     a(name="ex--json")
-  ssh-pre(language="json" label="package.json").
+  ssh-pre(language="json" label="package.json" :dark="$waveui.theme === 'dark'").
     {
       "name": "",
       "version": "1.0.0",
@@ -448,7 +502,7 @@
   h3
     a(href="#ex--php") PHP
     a(name="ex--php")
-  ssh-pre(language="php" label="index.php").
+  ssh-pre(language="php" label="index.php" :dark="$waveui.theme === 'dark'").
     &lt;?php
 
     require_once __DIR__ . '/../autoload.php';
@@ -513,7 +567,7 @@
     a(href="#ex--sql") SQL
     a(name="ex--sql")
 
-  ssh-pre(language="sql" label="SQL").
+  ssh-pre(language="sql" label="SQL" :dark="$waveui.theme === 'dark'").
     SET SQL_MODE = "NO_AUTO_VALUE_ON_ZERO";
     SET time_zone = "+00:00";
 
@@ -569,7 +623,14 @@
   h2
     a(href="#dark-mode") Dark mode
     a(name="dark-mode")
-  p The #[span.code dark] option switches the color theme to dark. Here is an example of Javascript code in dark mode.
+  p The #[span.code dark] option switches the color theme to dark.
+  p.text-bold(v-if="$waveui.theme === 'dark'").
+    You are already viewing this page in dark mode, but you could try switching to light mode
+    to view the difference.
+  p(v-else).
+    The #[span.code dark] option switches the color theme to dark. Here is an example of Javascript
+    code in dark mode.
+
 
   ssh-pre(language="js" label="Javascript" dark).
     /**
@@ -617,7 +678,7 @@
     You can easily create your own color theme, as the highlighting is already done,
     you just need to override the colors of these classes in your css:
 
-  ssh-pre(language="css" label="CSS").
+  ssh-pre(language="css" label="CSS" :dark="$waveui.theme === 'dark'").
     .txt {color: #333;}
     .comment {color: #aaa;}
     .quote {color: #c11;}
@@ -665,15 +726,18 @@
 </template>
 
 <script>
-import SimpleSyntaxHighlighter from '@/components/simple-syntax-highlighter.vue'
+import SshPre from '@/components/simple-syntax-highlighter.vue'
 import Highlight from '@/components/highlight-message.vue'
 
 export default {
-  components: { sshPre: SimpleSyntaxHighlighter, Highlight }
+  components: { SshPre, Highlight }
 }
 </script>
 
 <style lang="scss">
+.theme-switch {z-index: 100;}
+.checklist {list-style-type: none;}
+
 @media screen and (max-width: 640px) {
   .responsive-block {
     flex-direction: column;
