@@ -1,26 +1,28 @@
 import { defineConfig } from 'vite'
+import Vue from '@vitejs/plugin-vue'
 import { resolve } from 'path'
 import Delete from 'rollup-plugin-delete'
-import { createVuePlugin } from 'vite-plugin-vue2'
+import autoprefixer from 'autoprefixer'
 
 const build = process.env.BUNDLE ? {
   lib: {
     entry: resolve(__dirname, '/src/components/simple-syntax-highlighter.vue'),
     name: 'sshpre',
-    fileName: 'sshpre',
     formats: ['es', 'umd', 'cjs']
   },
   rollupOptions: {
     plugins: [
-      Delete({ targets: ['dist/*.{ico,html}'], hook: 'generateBundle' })
+      // Rollup also copies the files in the public folder.
+      // @todo: find a way to prevent adding them at all.
+      Delete({ targets: ['dist/*.ico'], hook: 'generateBundle' })
     ],
     // Make sure to externalize deps that shouldn't be bundled into library.
     external: ['vue'],
     output: {
       // Provide global variables to use in the UMD build for externalized deps.
-      globals: {
-        vue: 'Vue'
-      }
+      globals: { vue: 'Vue' },
+      entryFileNames: 'sshpre.[format].js',
+      chunkFileNames: '[name].js'
     }
   }
 } : {
@@ -29,14 +31,19 @@ const build = process.env.BUNDLE ? {
 
 export default defineConfig({
   plugins: [
-    createVuePlugin({
-      vueTemplateOptions: {
+    Vue({
+      template: {
         compilerOptions: {
           whitespace: 'preserve'
         }
       }
     })
-  ], // https://vitejs.dev/config/
+  ],
+  css: {
+    postcss: {
+      plugins: [autoprefixer]
+    }
+  },
   resolve: {
     alias: {
       '@': resolve(__dirname, '/src')
